reorganize JoinGame directory
