+++ client/src/JoinGame/components/Teams.js
