+++ client/src/JoinGame/components/categories/Categories.css
@@ -1,4 +1,4 @@
-@import "../common/Styles.css";
+@import "../../../common/Styles.css";
 
 .category-dropdown {
   width: 30%;
